Enable the use of the OverlayFS for the LiveOS root filesystem.
Integrate the option to use an OverlayFS as the root filesystem
into the 90dmsquash-live module for testing purposes.

Tested primarily with transient, in-RAM overlay boots on vfat-
formatted Live USB device, with persistent overlay directories
on ext4-formatted Live USB devices, and with embedded, persistent
overlay directories on vfat-formatted devices. (Persistent overlay
directories on a vfat-formatted device must be in an embedded
filesystem that supports the creation of trusted.* extended attributes,
and must provide valid d_type in readdir responses.)  The
rd.live.overlay.readonly option to allow a persistent overlayfs to be
mounted read only through a higher level transient overlay directory
has been implemented through the multiple lower layers feature of
OverlayFS.

This implementation currently requires the enforcing=0 kernel command
line option.  Further SELinux tuning may change this limitation.

To permit easier comparison of Device-mapper vs. OverlayFS performance,
the default transient dm overlay size has been adjusted up to 32 GiB.
A transient dm overlay is a sparse file in memory, so this does not
consume more RAM for legacy applications.  It does permit a user to
use all of the available root filesystem storage, and fails gently
when it is consumed, as the available free root filesystem storage on
a typical LiveOS build is under a few GiB.  Thus, when booted on other-
than-small RAM systems, the transient dm overlay should not overflow.

OverlayFS offers the potential to use all of the available free RAM,
or all of the available free disc storage on non-vfat-devices,
in its overlay, even beyond the root filesystem available space,
because the OverlayFS root filesystem is a union of directories on
two different partitions.